Bridge bare-scalar return widening across more sites
- Defer-return path now appends `;` after the return value so its
  temporaries drop before wrapped locals do; `return (*x.borrow()...)`
  inside the defer execution block was extending Refs past their locals.
- Treat short-decl wrapped locals as wrapped in tail-return fallback by
  trusting go/types for body-declared variables that aren't in VarTable.
  The Copy-scalar exemption was wrong because the default short-decl
  path wraps even bare-scalar locals.
- Recognize calls through function-typed values (closures, function
  params/fields) as bare-scalar-returning when their signature says so.
  The generated function-value box already uses GoReturnTypeToRust for
  results, so the call site must agree.
- Make assignment-RHS lowering (`writeArraySliceElementAssignmentValue`)
  skip the .borrow().as_ref().unwrap() unwrap for calls whose result is
  already bare.
- Mark CallExpr as bare in isExpressionResultBare when the call returns a
  bare scalar, so downstream wrapping paths stop wrapping it.
- Cast range usize indexes to i32 when binary-compared against any int
  expression (not just integer constants) — bare-scalar-returning calls
  like `limit(values)` are the new motivating case.
- Update hand-written stdlib stub method signatures (time.Time.Unix,
  GoTimer.stop, GoAtomicPointer.compare_and_swap, GoRegexp.match_string,
  GoReflectType.num_field, etc.) to return bare scalars so they match
  the widened signature callers now expect.

+++ tests/interface_slice_index_arg/main.rs
@@ -76,6 +76,6 @@ fn main() {
         if (i as i32) == (((*specs.borrow()).as_ref().map(|__v| __v.len()).unwrap_or(0) as i32) - (1 as i32) as i32) {
         continue
     }
-        println!("{}", format!("{}", pair_o_k(s.clone(), (*specs.borrow().as_ref().unwrap())[(i + 1) as usize].clone().clone())));
+        println!("{}", format!("{}", pair_o_k(s.clone(), (*specs.borrow().as_ref().unwrap())[(i as i32 + 1) as usize].clone().clone())));
     } }
 }
